requirements.txt
pillow
pytesseract
opencv-python
numpy
textdistance
ImageHash
pyinstaller